requirements.txt
streamlit
pandas
numpy
gspread
google-auth-oauthlib
google-api-python-client
oauth2client
openpyxl